deadsea_scrolls/requirements.txt
albumentations==1.3.0
imageio==2.28.1
imgaug==0.4.0
numpy==1.21.4
pandas==2.0.1
Pillow==9.5.0
ultralytics==8.0.43
optuna==3.1.1
openpyxl==3.1.2
opencv-python~=4.7.0.72
joblib~=1.2.0
tqdm~=4.65.0
